More refactoring updates
